+++ src/rx-database.js
@@ -295,11 +295,7 @@ export class RxDatabase {
             // collection already exists with different schema, check if it has documents
             const pouch = this._spawnPouchDB(args.name, args.schema.version, args.pouchSettings);
             const oneDoc = await pouch.find({
-                selector: {
-                    language: {
-                        $ne: 'query'
-                    }
-                },
+                selector: {},
                 limit: 1
             });
             if (oneDoc.docs.length !== 0) {